+++ src/tasks/CssFileHandler.php
@@ -34,12 +34,27 @@ class CssFileHandler extends BaseFileHandler
     }
 
     /**
+     * If css section does not exists in unglue config file return count 0 in order
+     * to make sure handler is not started with empty section.
+     * 
+     * @return integer
+     */
+    public function count()
+    {
+        if (!$this->getCssFilesFromConfig()) {
+            return 0;
+        }
+
+        return parent::count();
+    }
+
+    /**
      * {@inheritDoc}
      */
     public function handleUpload()
     {
         $mainFiles = [];
-        foreach ($this->config->getHasUnglueConfigSection('css', []) as $scss) {
+        foreach ($this->getCssFilesFromConfig() as $scss) {
             $mainFiles[] = $this->config->getUnglueConfigFolderPath($scss);
         }
 
@@ -54,5 +69,16 @@ class CssFileHandler extends BaseFileHandler
         }
 
         return false;
+    }
+
+    /**
+     * Return an array of files defined in unglue config "css" section.
+     *
+     * @return array An array where the value is the path to the css file from a relative view of the file.
+     * @since 1.0.1
+     */
+    protected function getCssFilesFromConfig()
+    {
+        return $this->config->getHasUnglueConfigSection('css', []);
     }
 }